requirements.txt
# Editable install with no version control (apriltag==0.0.16)
# Sorry anyone else | pip install apriltag will work on non apple silicon
-e /Users/zynh/Downloads/apriltag-0.0.16/python
numpy==1.23.4
opencv-python==4.6.0.66
pynetworktables==2021.0.0
python-dotenv==0.21.0
